% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{best_models}
\alias{best_models}
\title{A list of the best predictive model for FIT's perforemance per threhsolds pair (fold-change and q-value)}
\format{An object of class \code{list} of length 117.}
\usage{
best_models
}
\description{
A list of the best predictive model for FIT's perforemance per threhsolds pair (fold-change and q-value)
}
\keyword{datasets}
