% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{GetRefData}
\alias{GetRefData}
\title{Download Reference Data}
\usage{
GetRefData()
}
\value{
A data.frame containing the following columns:
\itemize{
  \item{\strong{Disease} - Disease name.}
  \item{\strong{DataType} - <Technology>_<RF/ST> - \strong{Technology} can be either RNAseq or Microarrays.
         \strong{RF/ST} can either be \strong{RF} (Reference) denoting a cross-species pairing (CSP) where the human and mouse datasets were directly 
         contrasted to one another in a publication authored by the researchers who had generated at least one of the datasets, 
         or \strong{ST} (Standard) denoting a CSP which consisted of human datasets and corresponding datasets of a mouse model of the 
         human disease in a separate study.}
  \item{\strong{CSP_ID} - ID of the CSP.} 
  \item{\strong{MM.Entrez, HS.Entrez} - Gene IDs.}
  \item{\strong{FC.HS, FC.MM} - Mouse (MM) and human (HS) fold-change for the CSP.}
  \item{\strong{EffSize.HS, EffSize.MM} - Mouse and human effect sizes for the CSP. 
                Z-tests for microarrays, estimated fold-change, as computed in Sleuth for RNA-seq.}
  \item{\strong{qval.HS, qval.MM} - Mouse (MM) and human (HS) q-values for the CSP.}
  }
}
\description{
This function allows access to the latest reference data used by the FIT model
}
