% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{GetSampleData}
\alias{GetSampleData}
\title{Download sampel data}
\usage{
GetSampleData(DataType)
}
\arguments{
\item{DataType}{Either "microarray" or "rnaseq", depending on the technology by which the data was assayed.}
}
\description{
This function allows download of sample data that can be used by FIT, in CSV format.
The file will be saved in the name "SampleRNAseq.csv" or "SampleMicroarray.csv" depending on the DataType chosen.
}
