% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HS_MM_Symbol_Entrez}
\alias{HS_MM_Symbol_Entrez}
\title{A data.frame containing the Entrez IDs of human and mouse orthologs}
\format{An object of class \code{data.frame} with 16425 rows and 4 columns.}
\usage{
HS_MM_Symbol_Entrez
}
\description{
A data.frame containing the Entrez IDs of human and mouse orthologs
}
\keyword{datasets}
