% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MGD_orthologs}
\alias{MGD_orthologs}
\title{A data.frame containing the Entrez IDs of human and mouse orthologs}
\format{An object of class \code{data.frame} with 16491 rows and 2 columns.}
\usage{
MGD_orthologs
}
\description{
A data.frame containing the Entrez IDs of human and mouse orthologs
}
\keyword{datasets}
