% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{microarray_sample}
\alias{microarray_sample}
\title{Microarray sample data}
\format{An object of class \code{data.frame} with 11545 rows and 10 columns.}
\usage{
microarray_sample
}
\description{
Microarray sample data
}
\keyword{datasets}
