% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MM_Entrez_symbol_desc}
\alias{MM_Entrez_symbol_desc}
\title{A data.frame containing the symbols of human and mouse orthologs and a description per gene}
\format{An object of class \code{data.frame} with 21025 rows and 3 columns.}
\usage{
MM_Entrez_symbol_desc
}
\description{
A data.frame containing the symbols of human and mouse orthologs and a description per gene
}
\keyword{datasets}
