% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pca_rotations}
\alias{pca_rotations}
\title{The rotations needed to convert a vector of fold-changes to a dot on the PC space of all cross-species pairs in the training data}
\format{An object of class \code{matrix} with 4957 rows and 137 columns.}
\usage{
pca_rotations
}
\description{
The rotations needed to convert a vector of fold-changes to a dot on the PC space of all cross-species pairs in the training data
}
\keyword{datasets}
