% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{PreProcess}
\alias{PreProcess}
\title{Pre-processing of mouse data for predictions by FIT}
\usage{
PreProcess(MouseData, DataType)
}
\arguments{
\item{MouseData}{The mouse data.}

\item{DataType}{'microarray' or 'rnaseq'}
}
\description{
First, the genes are filtered only to those that have slopes in the model.
}
\details{
Pre-processing for microarrays includes 4 additional steps:
a) Computing fold-change per gene
b) Computing Z-scores per gene
c) Computing Z-tests per gene
d) Merging human orthologs
}
