% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Psoriasis_classifier_neg}
\alias{Psoriasis_classifier_neg}
\title{Microarray sample data to run in the function RunClassifier - that produces a negative result (FIT is unlikely to improve the data, for FC=0.15 qval=0.1)}
\format{An object of class \code{data.frame} with 11545 rows and 10 columns.}
\usage{
Psoriasis_classifier_neg
}
\description{
Microarray sample data to run in the function RunClassifier - that produces a negative result (FIT is unlikely to improve the data, for FC=0.15 qval=0.1)
}
\keyword{datasets}
