% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RNAseq_sample}
\alias{RNAseq_sample}
\title{RNAseq sample data}
\format{An object of class \code{data.frame} with 14916 rows and 2 columns.}
\usage{
RNAseq_sample
}
\description{
RNAseq sample data
}
\keyword{datasets}
