% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{RunClassifier}
\alias{RunClassifier}
\title{Run FIT improvement prediction classifier}
\usage{
RunClassifier(MouseData = NULL, MouseFile = NULL, DataType = NULL,
  qval = 0.1, FC = 0.15, verbose = F)
}
\arguments{
\item{MouseData}{The pre-processed mouse dataset, or NULL in case MouseFIle ise given}

\item{MouseFile}{File name that includes the mouse data (log expression per gene for all disease and control sampels), in CSV format. NULL in case MouseData is given}

\item{DataType}{Either "microarray" or "rnaseq", depending on the technology by which the data was assayed. NULL if MouseData is given}

\item{qval}{the q-value cuttoff the user will use to interpret FIT's predictions. (default= 0.1)}

\item{FC}{the fold-change cuttoff the user will use to interpret FIT's predictions, given as fraction from the top. For example, 
0.15 denotes the top 15\% of genes with highest fold-change. (default= 0.15)}
}
\description{
The SVM classifier predicts whether FIT will be able to improve a specific mouse data.
}
