% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{ShowClassifierPerformance}
\alias{ShowClassifierPerformance}
\title{ShowClassifierPerformance}
\usage{
ShowClassifierPerformance()
}
\description{
Show performance of SVM classifier (as an image, taken from the paper)
}
