% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{slopes_per_gene_V2.0}
\alias{slopes_per_gene_V2.0}
\title{The slopes derived from the linear regression FIT computed from human and mouse data, per gene. This is used for the prediciton of human effect size values.}
\format{An object of class \code{array} of length 15212.}
\usage{
slopes_per_gene_V2.0
}
\description{
The slopes derived from the linear regression FIT computed from human and mouse data, per gene. This is used for the prediciton of human effect size values.
}
\keyword{datasets}
