% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StaphInfection_classifier_pos}
\alias{StaphInfection_classifier_pos}
\title{Microarray sample data to run in the function RunClassifier - that produces a positive result (FIT is likely to improve the data, for FC=0.15 qval=0.1)}
\format{An object of class \code{data.frame} with 11545 rows and 20 columns.}
\usage{
StaphInfection_classifier_pos
}
\description{
Microarray sample data to run in the function RunClassifier - that produces a positive result (FIT is likely to improve the data, for FC=0.15 qval=0.1)
}
\keyword{datasets}
