% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{CheckFile}
\alias{CheckFile}
\title{Check input file}
\usage{
CheckFile(MouseFile)
}
\arguments{
\item{MouseFile}{The mouse data file}
}
\description{
This function validates that the input file exists and contain a valid number of genes in the first row
}
