% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{CheckFormat}
\alias{CheckFormat}
\title{Check input file format}
\usage{
CheckFormat(MouseData)
}
\arguments{
\item{MouseData}{The mouse data}
}
\description{
This function validates the format of the input mouse gene expression file is correct for process by FIT.
It will output an error message in 3 cases: 
(a) Less than 80% of the gene IDs are Entrez IDs; 
(b) Sample names don't start with "c_" or "d_"
(c) There are at least 3 disease samples and 3 control samples
(d) The data i snot log-transformed (the range of values are either <0 or >100)
}
