% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{FIT}
\alias{FIT}
\title{Run FIT pipeline}
\usage{
FIT(MouseFile)
}
\arguments{
\item{NewMouse_df}{The pre-processed mouse dataset}
}
\description{
This function runs the whole FIT pipeline: checks input file format, pre-processes data and computes predictions.
}
