% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{PreProcess}
\alias{PreProcess}
\title{Pre-processing of mouse data for predictions by FIT (for microarray data only)}
\usage{
PreProcess(MouseData)
}
\arguments{
\item{MouseData}{The mouse data.}
}
\description{
Pre-processing includes 4 steps:
a) Computing fold-change per gene
b) Computing Z-scores per gene
c) Computing Z-tests per gene
d) Merging human orthologs
}
