% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{RunClassifier}
\alias{RunClassifier}
\title{Run FIT improvement prediction classifier}
\usage{
RunClassifier(MouseFile, qval = 0.1, FC = 0.15)
}
\arguments{
\item{MouseFile}{File name that includes the mouse data (log expression per gene for all disease and control sampels), in CSV format}

\item{qval}{the q-value cuttoff the user will use to interpret FIT's predictions. (default= 0.1)}

\item{FC}{the fold-change cuttoff the user will use to interpret FIT's predictions, given as fraction from the top. For example, 
0.15 denotes the top 15\% of genes with highest fold-change. (default= 0.15)}
}
\description{
The SVM classifier predicts whether FIT will be able to improve a specific mouse data
}
